change int to double on delay not pokeballs

+++ PokemonGoGUI/Settings.cs
@@ -24,7 +24,7 @@ namespace PokemonGoGUI
         public string TimeZone { get; set; }
         public string POSIX { get; set; }
 
-        public int DisableCatchDelay { get; set; }
+        public double DisableCatchDelay { get; set; }
         public bool SpinGyms { get; set; }
         public string GroupName { get; set; }
         public string AccountName { get; set; }
